requirements.txt
streamlit
openai
pandas
python-dotenv
faiss-cpu
langchain
langchain-community
tiktoken
reportlab
